feat: alert delete artwork

+++ App/src/component/admin/artwork_list.jsx
@@ -7,7 +7,6 @@ import { useNavigate, Link } from "react-router-dom";
 
 export default function ArtworkList({ sendDataAddButton }) {
   const [art, setArt] = useState([]);
-  // const [id, setId] = useState();
   const navigate = useNavigate();
 
   async function getArtworkList() {
@@ -24,17 +23,19 @@ export default function ArtworkList({ sendDataAddButton }) {
   }, []);
 
   const handleAddArtworkList = () => {
-    // setMode('add');
     navigate("/admin/artwork/artwork-add");
   };
 
   const handleDeleteArtwork = async (id) => {
-    try {
-      await axios.delete(`http://localhost:3000/artwork_list/${id}`);
-      // Refresh the artwork list after deletion
-      getArtworkList();
-    } catch (error) {
-      console.log(error);
+    const shouldDelete = window.confirm("Are you sure you want to delete this Artwork?");
+
+    if (shouldDelete) {
+      try {
+        await axios.delete(`http://localhost:3000/artwork_list/${id}`);
+        getArtworkList();
+      } catch (error) {
+        console.log(error);
+      }
     }
   };
